requirements.txt
fastapi
uvicorn
jinja2
python-multipart
pandas
openpyxl
pdfplumber
pymupdf
pytesseract
Pillow
python-slugify
slowapi